refactor : email 중복 거부

+++ pages/api/auth/register/check/email.ts
@@ -9,38 +9,48 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   const { email, token, type }: HelpForm = req.body;
   console.log(email, token);
   if (email && !token) {
-    const payload = Math.floor(10000 + Math.random() * 1000000) + "";
-    const mailOptions = {
-      from: process.env.MAIL_ID,
-      to: email,
-      subject: "이메일 인증",
-      text: `인증코드 : ${payload}`,
-    };
-    const result = await smtpTransport.sendMail(mailOptions, (error, responses) => {
-      if (error) {
-        console.log(error);
-        return null;
-      } else {
-        console.log(responses);
-        return null;
-      }
-    });
-    smtpTransport.close();
-
-    await client.certification.deleteMany({
-      where: { email },
-    });
-
-    await client.certification.create({
-      data: {
-        number: payload,
+    const user = await client.user.findFirst({
+      where: {
+        type,
         email,
       },
     });
-    console.log(payload);
-    return res.status(200).json({
-      ok: true,
-    });
+    const payload = Math.floor(10000 + Math.random() * 1000000) + "";
+    if (user) {
+      return res.status(403).send("중복된 이메일입니다");
+    } else {
+      const mailOptions = {
+        from: process.env.MAIL_ID,
+        to: email,
+        subject: "이메일 인증",
+        text: `인증코드 : ${payload}`,
+      };
+      const result = await smtpTransport.sendMail(mailOptions, (error, responses) => {
+        if (error) {
+          console.log(error);
+          return null;
+        } else {
+          console.log(responses);
+          return null;
+        }
+      });
+      smtpTransport.close();
+
+      await client.certification.deleteMany({
+        where: { email },
+      });
+
+      await client.certification.create({
+        data: {
+          number: payload,
+          email,
+        },
+      });
+      console.log(payload);
+      return res.status(200).json({
+        ok: true,
+      });
+    }
   }
   if (email && token) {
     const FindToken = await client.certification.deleteMany({
